requirements.txt
flask
python-dotenv
tweepy
pandas
matplotlib
wordcloud
nltk